+++ ChangeLog.md
@@ -3,6 +3,8 @@ Measuring API change log
 
 ## ?.?.? / ????-??-??
 
+* Catch exceptions from methods and report them in outcome - @thekid
+
 ## 0.2.0 / 2015-01-10
 
 * Made available via Composer - @thekid