+++ ChangeLog.md
@@ -3,6 +3,10 @@ Measuring API change log
 
 ## ?.?.? / ????-??-??
 
+* Added support for `@values` annotation which can produce permutations
+  to be used for parametrized measurements. Supports arrays or references
+  to static class methods of measurable class.
+  (@thekid)
 * Catch exceptions from methods and report them in outcome - @thekid
 
 ## 0.2.0 / 2015-01-10